DNA_Sequencing/requirements.txt
flask>=2.0.0
biopython>=1.79
reportlab>=3.6.0
werkzeug>=2.0.0
numpy>=1.21.0
